move autoRequire script. and update autoRequire flag description
